Seperated JS functions into seperate files
Seperated JS functions into seperate files
